requirements.txt
streamlit==1.31.1
google-generativeai==0.8.3
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.30
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.5
langsmith==0.2.11
chromadb==0.4.22
tiktoken==0.6.0
pypdf==4.0.0
python-dotenv==1.0.1